requirements.txt
asgiref==3.2.10
bleach==3.1.5
dj-database-url==0.5.0
Django==3.0.8
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-taggit==1.3.0
django-tinymce==3.0.2
gunicorn==20.0.4
importlib-metadata==1.7.0
install==1.3.3
Markdown==3.2.2
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
Unipath==1.1
webencodings==0.5.1
whitenoise==5.1.0
zipp==3.1.0
